The FG scan is started by D7ATP.
- In case of broadcast request, the responder is allowed to postpone
the FG scan until the response period expires. A Tc timer is started
for that purpose only. For unicast request, the FG scan is started
once the frame is verified by D7ATP even if no response nor ACK is
expected.

- To be aligned with the DLL behavior, D7ANP restores the state to
D7ANP_STATE_FOREGROUND_SCAN after a packet is transmitted if the FG
scan was active prior to the transmission.